fix a bug in no music function. moved the bias from 519 to 81 changed Servo class to VarSpeedServo Computing state is not used anymore

+++ project/StateManager.h
@@ -11,6 +11,6 @@ bool is_high_pitch(int value);
 
 bool is_computing(int value);
 
-bool is_music_off(int value, boolean no_update);
+bool is_music_off(float value, boolean no_update);
 
 #endif
